refactor: 로또번호 개수 검증 로직 Lottery로 이동
