+++ src/main/java/lotto/domain/LottoNumberValidator.java
@@ -2,19 +2,13 @@ package lotto.domain;
 
 import java.util.Set;
 
-import static lotto.domain.LottoNumberGenerator.*;
+import static lotto.domain.LottoNumberGenerator.LOTTO_NUMBER_LOWER_LIMIT;
+import static lotto.domain.LottoNumberGenerator.LOTTO_NUMBER_UPPER_LIMIT;
 
 public class LottoNumberValidator {
     public static void validateLottoNumbers(Set<Integer> lottoNumbers) {
-        validateLottoNumberSize(lottoNumbers);
         for (Integer lottoNumber : lottoNumbers) {
             validateLottoNumberRange(lottoNumber);
-        }
-    }
-
-    private static void validateLottoNumberSize(Set<Integer> lottoNumbers) {
-        if (lottoNumbers == null || lottoNumbers.size() != LOTTO_NUMBERS_SIZE) {
-            throw new IllegalArgumentException(String.format("로또번호는 %d개 이어야 합니다", LOTTO_NUMBERS_SIZE));
         }
     }
