+++ src/test/java/lotto/LottoNumberValidatorTest.java
@@ -8,17 +8,6 @@ import java.util.Set;
 
 public class LottoNumberValidatorTest {
     @Test
-    void 로또번호가_6개가_아니면_예외_발생() {
-        // given
-        Set<Integer> lottoNumbers = Set.of(1, 2, 3, 4, 5);
-
-        // when, then
-        Assertions.assertThatExceptionOfType(IllegalArgumentException.class)
-                .isThrownBy(() -> LottoNumberValidator.validateLottoNumbers(lottoNumbers))
-                .withMessageMatching("로또번호는 \\d+개 이어야 합니다");
-    }
-
-    @Test
     void 로또번호가_정해진_범위내의_숫자가_아니면_예외_발생() {
         // given
         Set<Integer> lottoNumbers = Set.of(46, 1, 2, 3, 4, 5);
